refactor(codegen): extrai compile/class.rs de func.rs
Move synthesize_class_fns + validate_abstract_method_implementations
+ scan_this_* + make_field_init_stmt + count_super_calls_in_top_level
+ weave_initializers + is_super_call_stmt + synth_abstract_stub_body
+ this_param + class_*_name (601 linhas) para
crates/rts-codegen/src/codegen/lower/compile/class.rs.

Os helpers `class_*_name` agora moram em compile/class e func.rs faz
re-export pra preservar callers em expressions/*.

func.rs: 2070 -> 1453 linhas. Suite verde (11/11 crates).
