+++ crates/rts-codegen/src/codegen/lower/func.rs
@@ -11,13 +11,11 @@ use cranelift_codegen::ir::{AbiParam, InstBuilder, Signature, types as cl};
 use cranelift_codegen::isa::CallConv;
 use cranelift_frontend::{FunctionBuilder, FunctionBuilderContext};
 use cranelift_module::{Linkage, Module};
-use swc_ecma_ast::{Callee, Decl, Expr, ForHead, Lit, MemberProp, Pat, Stmt};
+use swc_ecma_ast::{Decl, ForHead, Pat, Stmt};
 
 use crate::parser::ast::{
-    ClassDecl, ClassMember, FunctionDecl, Item, MemberModifiers, MethodRole, Parameter, Program,
-    RawStmt, Statement,
+    ClassDecl, FunctionDecl, Item, Program, Statement,
 };
-use crate::parser::span::Span;
 
 use super::analysis::address_taken::collect_address_taken_fns;
 use super::analysis::captures::extract_class_owner;
@@ -33,10 +31,19 @@ use super::passes::object_methods::desugar_object_methods;
 use super::passes::parallelism::{
     array_methods_pass, lift_inline_arrows_in_array_methods, purity_pass, reduce_pass,
 };
+use super::compile::class::{
+    class_init_name, synthesize_class_fns, validate_abstract_method_implementations,
+};
 use super::passes::static_fields::expand_static_fields;
 use super::passes::this_arrow::lift_arrow_callbacks;
 use super::ctx::{ClassMeta, FnCtx, GlobalVar, UserFnAbi, ValTy};
 use super::statements::lower_stmt;
+
+// Re-export para callers externos (expressions/*) que ainda referenciam
+// `lower::func::class_*_name`.
+pub(crate) use super::compile::class::{
+    class_getter_name, class_setter_name, class_static_method_name,
+};
 
 const RUNTIME_MAIN_SYMBOL: &str = crate::abi::symbols::ENTRY_POINT;
 
@@ -1451,616 +1458,4 @@ fn compile_main_entry_shim(
 
     Ok(())
 }
-
-// ── Class lowering ────────────────────────────────────────────────────────
-
-/// Sintetiza os FunctionDecl para uma classe: constructor + cada metodo
-/// vira uma funcao independente que recebe `this` como primeiro parametro.
-/// Retorna o ClassMeta usado pelo codegen para resolver `new` e dispatch.
-/// Verifica que toda classe concreta implementa os métodos abstract
-/// herdados de seus ancestrais. Coleta o conjunto de abstracts da
-/// hierarquia, subtrai os métodos concretos efetivamente declarados
-/// e exige conjunto vazio.
-fn validate_abstract_method_implementations(classes: &HashMap<String, ClassMeta>) -> Result<()> {
-    for (name, meta) in classes {
-        if meta.is_abstract {
-            continue; // abstract classes podem deixar abstracts pendentes
-        }
-
-        // Acumula abstracts da hierarquia.
-        let mut required: std::collections::HashSet<String> = std::collections::HashSet::new();
-        let mut cur = Some(name.clone());
-        while let Some(c) = cur {
-            if let Some(m) = classes.get(&c) {
-                for am in &m.abstract_methods {
-                    required.insert(am.clone());
-                }
-                cur = m.super_class.clone();
-            } else {
-                break;
-            }
-        }
-
-        // Subtrai métodos concretos providos pela classe ou ancestrais.
-        let mut cur = Some(name.clone());
-        while let Some(c) = cur {
-            if let Some(m) = classes.get(&c) {
-                for method in &m.methods {
-                    if !m.abstract_methods.contains(method) {
-                        required.remove(method);
-                    }
-                }
-                cur = m.super_class.clone();
-            } else {
-                break;
-            }
-        }
-
-        if !required.is_empty() {
-            let mut missing: Vec<&str> = required.iter().map(|s| s.as_str()).collect();
-            missing.sort();
-            return Err(anyhow!(
-                "classe concreta `{name}` nao implementa metodo(s) abstract: {}",
-                missing.join(", ")
-            ));
-        }
-    }
-    Ok(())
-}
-
-/// (#nested-this) Escaneia um statement por padroes
-/// `this.X = { k: v, ... }` ou `this.X = { k: { ... } }` e popula
-/// field_obj_types / field_nested_obj_types. Inferencia simples
-/// baseada em tipos de literais (Str→Handle, Num→I64, Bool→Bool).
-fn scan_this_obj_assign(
-    s: &Statement,
-    field_obj_types: &mut HashMap<String, HashMap<String, ValTy>>,
-    field_nested_obj_types: &mut HashMap<(String, String), HashMap<String, ValTy>>,
-) {
-    let Statement::Raw(rs) = s;
-    let Some(stmt) = rs.stmt.as_ref() else { return };
-    scan_this_stmt(stmt, field_obj_types, field_nested_obj_types);
-}
-
-fn scan_this_stmt(
-    stmt: &swc_ecma_ast::Stmt,
-    field_obj_types: &mut HashMap<String, HashMap<String, ValTy>>,
-    field_nested_obj_types: &mut HashMap<(String, String), HashMap<String, ValTy>>,
-) {
-    use swc_ecma_ast::*;
-    match stmt {
-        Stmt::Expr(e) => scan_this_expr(&e.expr, field_obj_types, field_nested_obj_types),
-        Stmt::Block(b) => {
-            for s in &b.stmts {
-                scan_this_stmt(s, field_obj_types, field_nested_obj_types);
-            }
-        }
-        Stmt::If(i) => {
-            scan_this_stmt(&i.cons, field_obj_types, field_nested_obj_types);
-            if let Some(alt) = &i.alt {
-                scan_this_stmt(alt, field_obj_types, field_nested_obj_types);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn scan_this_expr(
-    e: &swc_ecma_ast::Expr,
-    field_obj_types: &mut HashMap<String, HashMap<String, ValTy>>,
-    field_nested_obj_types: &mut HashMap<(String, String), HashMap<String, ValTy>>,
-) {
-    use swc_ecma_ast::*;
-    let Expr::Assign(a) = e else { return };
-    if !matches!(a.op, AssignOp::Assign) {
-        return;
-    }
-    // LHS: this.X
-    let AssignTarget::Simple(SimpleAssignTarget::Member(m)) = &a.left else {
-        return;
-    };
-    if !matches!(m.obj.as_ref(), Expr::This(_)) {
-        return;
-    }
-    let MemberProp::Ident(field_id) = &m.prop else {
-        return;
-    };
-    let field_name = field_id.sym.as_str().to_string();
-    // RHS: object literal
-    let Expr::Object(obj) = a.right.as_ref() else {
-        return;
-    };
-    let mut fts: HashMap<String, ValTy> = HashMap::new();
-    for prop in &obj.props {
-        if let PropOrSpread::Prop(p) = prop {
-            if let Prop::KeyValue(kv) = p.as_ref() {
-                let key = match &kv.key {
-                    PropName::Ident(i) => i.sym.as_str().to_string(),
-                    PropName::Str(s) => s.value.to_string_lossy().to_string(),
-                    _ => continue,
-                };
-                match kv.value.as_ref() {
-                    Expr::Lit(Lit::Str(_)) => {
-                        fts.insert(key, ValTy::Handle);
-                    }
-                    Expr::Lit(Lit::Num(_)) => {
-                        fts.insert(key, ValTy::I64);
-                    }
-                    Expr::Lit(Lit::Bool(_)) => {
-                        fts.insert(key, ValTy::Bool);
-                    }
-                    Expr::Object(sub) => {
-                        fts.insert(key.clone(), ValTy::Handle);
-                        let mut sub_fts: HashMap<String, ValTy> = HashMap::new();
-                        for sp in &sub.props {
-                            if let PropOrSpread::Prop(spx) = sp {
-                                if let Prop::KeyValue(skv) = spx.as_ref() {
-                                    let sk = match &skv.key {
-                                        PropName::Ident(i) => i.sym.as_str().to_string(),
-                                        PropName::Str(s) => {
-                                            s.value.to_string_lossy().to_string()
-                                        }
-                                        _ => continue,
-                                    };
-                                    match skv.value.as_ref() {
-                                        Expr::Lit(Lit::Str(_)) => {
-                                            sub_fts.insert(sk, ValTy::Handle);
-                                        }
-                                        Expr::Lit(Lit::Num(_)) => {
-                                            sub_fts.insert(sk, ValTy::I64);
-                                        }
-                                        Expr::Lit(Lit::Bool(_)) => {
-                                            sub_fts.insert(sk, ValTy::Bool);
-                                        }
-                                        _ => {}
-                                    }
-                                }
-                            }
-                        }
-                        if !sub_fts.is_empty() {
-                            field_nested_obj_types
-                                .insert((field_name.clone(), key), sub_fts);
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-    if !fts.is_empty() {
-        field_obj_types.insert(field_name, fts);
-    }
-}
-
-fn synthesize_class_fns(class: &ClassDecl) -> (ClassMeta, Vec<FunctionDecl>) {
-    let mut methods: Vec<String> = Vec::new();
-    let mut getters: Vec<String> = Vec::new();
-    let mut setters: Vec<String> = Vec::new();
-    let mut static_methods: Vec<String> = Vec::new();
-    let mut static_fields: Vec<String> = Vec::new();
-    let mut fns: Vec<FunctionDecl> = Vec::new();
-    let mut field_types: HashMap<String, ValTy> = HashMap::new();
-    let mut field_class_names: HashMap<String, String> = HashMap::new();
-    let mut field_obj_types: HashMap<String, HashMap<String, ValTy>> = HashMap::new();
-    let mut field_nested_obj_types: HashMap<(String, String), HashMap<String, ValTy>> =
-        HashMap::new();
-    let mut readonly_fields: std::collections::HashSet<String> = std::collections::HashSet::new();
-    let mut abstract_methods: std::collections::HashSet<String> = std::collections::HashSet::new();
-    let mut member_visibility: std::collections::HashMap<String, crate::parser::ast::Visibility> =
-        std::collections::HashMap::new();
-    let mut has_constructor = false;
-
-    // Coleta initializers de instância (`x = expr`) na ordem declarada.
-    // Serão prependidos ao body do constructor (depois de `super()` se
-    // houver). Static props ficam fora — initializers static seriam
-    // tratados separadamente (não cobertos neste commit).
-    let init_stmts: Vec<Statement> = class
-        .members
-        .iter()
-        .filter_map(|m| match m {
-            ClassMember::Property(prop)
-                if !prop.modifiers.is_static && prop.initializer.is_some() =>
-            {
-                let init = prop.initializer.as_ref().unwrap().clone();
-                Some(make_field_init_stmt(&prop.name, init, prop.span))
-            }
-            _ => None,
-        })
-        .collect();
-
-    for member in &class.members {
-        match member {
-            ClassMember::Constructor(ctor) => {
-                has_constructor = true;
-                // (#303 parte 1) Detecta \`super(...); super(...)\` em
-                // sequencia direta (mesmo bloco/escopo top-level do
-                // constructor body). JS proibe — segundo super lanca
-                // ReferenceError. So' rejeita o caso linear obvio; em
-                // branches if/else mutuamente exclusivos (super em cons
-                // E em alt), passa silenciosamente — runtime check seria
-                // a fase 2 dessa issue.
-                if count_super_calls_in_top_level(&ctor.body) > 1 {
-                    // SyntaxError-like: rejeita o programa em compile time.
-                    // Usar eprintln + std::process::exit(1) imita o caminho
-                    // de erros existentes em outras validacoes (abstract,
-                    // visibility). Sem Result aqui pra nao espalhar
-                    // mudanca de tipo em todo synthesize_class_fns caller.
-                    eprintln!(
-                        "error: ReferenceError: super constructor may only be called once (em `{}`)",
-                        class.name
-                    );
-                    std::process::exit(1);
-                }
-                for p in &ctor.parameters {
-                    if let Some(ann) = p.type_annotation.as_deref() {
-                        field_types
-                            .entry(p.name.clone())
-                            .or_insert(ValTy::from_annotation(ann));
-                    }
-                }
-                let mut params = Vec::with_capacity(ctor.parameters.len() + 1);
-                params.push(this_param(ctor.span));
-                params.extend(ctor.parameters.iter().cloned());
-                // Body = [super() se houver no inicio] + initializers + user code.
-                // Detecta `super(...)` na primeira posição e injeta initializers
-                // logo depois (semântica TS: initializers rodam depois do
-                // super call).
-                let body = weave_initializers(&ctor.body, &init_stmts, class.super_class.is_some());
-                fns.push(FunctionDecl {
-                    name: class_init_name(&class.name),
-                    parameters: params,
-                    return_type: None,
-                    body,
-                    span: ctor.span,
-                    is_async: false,
-                });
-            }
-            ClassMember::Method(method) => {
-                // Visibility — registra apenas private/protected (public é default).
-                if let Some(v) = method.modifiers.visibility {
-                    if !matches!(v, crate::parser::ast::Visibility::Public) {
-                        member_visibility.insert(method.name.clone(), v);
-                    }
-                }
-                // Métodos abstract: gera um stub que faz `throw "abstract"`
-                // (na prática, retorna 0). O stub permite que o codegen
-                // resolva referências `__class_C_<m>` para checagem de
-                // assinatura, e o dispatch virtual roteia para a impl
-                // concreta da subclasse em runtime. Se chamado direto na
-                // base abstract (não deveria acontecer porque `new` é
-                // bloqueado), retorna o default da assinatura.
-                if method.modifiers.is_abstract {
-                    abstract_methods.insert(method.name.clone());
-                    if matches!(method.role, MethodRole::Method) {
-                        methods.push(method.name.clone());
-                    }
-                    let synth_name = match method.role {
-                        MethodRole::Getter => class_getter_name(&class.name, &method.name),
-                        MethodRole::Setter => class_setter_name(&class.name, &method.name),
-                        MethodRole::Method => class_method_name(&class.name, &method.name),
-                    };
-                    let mut params = Vec::with_capacity(method.parameters.len() + 1);
-                    params.push(this_param(method.span));
-                    params.extend(method.parameters.iter().cloned());
-                    // Body do stub: retorna o default do tipo declarado.
-                    // Se return_type é "void", body vazio basta. Caso
-                    // contrário, `return 0;` ou `return 0.0;`.
-                    let body = synth_abstract_stub_body(method.return_type.as_deref());
-                    fns.push(FunctionDecl {
-                        name: synth_name,
-                        parameters: params,
-                        return_type: method.return_type.clone(),
-                        body,
-                        span: method.span,
-                        is_async: false,
-                    });
-                    continue;
-                }
-                if method.modifiers.is_static {
-                    static_methods.push(method.name.clone());
-                    fns.push(FunctionDecl {
-                        name: class_static_method_name(&class.name, &method.name),
-                        parameters: method.parameters.clone(),
-                        return_type: method.return_type.clone(),
-                        body: method.body.clone(),
-                        span: method.span,
-                        is_async: false,
-                    });
-                } else {
-                    let synth_name = match method.role {
-                        MethodRole::Getter => {
-                            getters.push(method.name.clone());
-                            class_getter_name(&class.name, &method.name)
-                        }
-                        MethodRole::Setter => {
-                            setters.push(method.name.clone());
-                            class_setter_name(&class.name, &method.name)
-                        }
-                        MethodRole::Method => {
-                            methods.push(method.name.clone());
-                            class_method_name(&class.name, &method.name)
-                        }
-                    };
-                    let mut params = Vec::with_capacity(method.parameters.len() + 1);
-                    params.push(this_param(method.span));
-                    params.extend(method.parameters.iter().cloned());
-                    fns.push(FunctionDecl {
-                        name: synth_name,
-                        parameters: params,
-                        return_type: method.return_type.clone(),
-                        body: method.body.clone(),
-                        span: method.span,
-                        is_async: false,
-                    });
-                }
-            }
-            ClassMember::Property(prop) => {
-                // Visibility — registra apenas private/protected.
-                if let Some(v) = prop.modifiers.visibility {
-                    if !matches!(v, crate::parser::ast::Visibility::Public) {
-                        member_visibility.insert(prop.name.clone(), v);
-                    }
-                }
-                if prop.modifiers.is_static {
-                    static_fields.push(prop.name.clone());
-                } else {
-                    if let Some(ann) = prop.type_annotation.as_deref() {
-                        let ann = ann.trim();
-                        field_types.insert(prop.name.clone(), ValTy::from_annotation(ann));
-                        field_class_names.insert(prop.name.clone(), ann.to_string());
-                    }
-                    if prop.modifiers.readonly {
-                        readonly_fields.insert(prop.name.clone());
-                    }
-                    // Private fields sem anotação ainda precisam ser
-                    // detectáveis na hierarquia para validação de escopo.
-                    // Garantimos uma entrada em field_types (default I64).
-                    if prop.name.starts_with('#') && !field_types.contains_key(&prop.name) {
-                        field_types.insert(prop.name.clone(), ValTy::I64);
-                    }
-                }
-            }
-        }
-    }
-
-    // Se a classe não tem constructor explícito mas tem initializers,
-    // sintetizamos um ctor implícito que apenas executa-os. Para classes
-    // com `extends` mas sem ctor explícito, TS gera um pass-through
-    // `constructor(...args) { super(...args); }` — não suportamos rest
-    // args ainda (#58/#59), então damos erro claro nesse caso.
-    if !has_constructor && !init_stmts.is_empty() {
-        if class.super_class.is_some() {
-            // Sub sem ctor + extends + initializers: precisaria de
-            // `super(...args)` implícito. Por simplicidade do MVP, exija
-            // ctor explícito nesse caso.
-            // (Ainda emitimos o ctor implícito sem super — funciona se
-            // a classe pai não tem ctor com args.)
-        }
-        let init_only_body = weave_initializers(&[], &init_stmts, false);
-        fns.push(FunctionDecl {
-            name: class_init_name(&class.name),
-            parameters: vec![this_param(class.span)],
-            return_type: None,
-            body: init_only_body,
-            span: class.span,
-            is_async: false,
-        });
-        has_constructor = true;
-    }
-
-    // (#nested-this) Escaneia o body do constructor (e initializers) por
-    // `this.X = { sub: { ... } }` ou `this.X = { ... }` e popula
-    // field_obj_types / field_nested_obj_types. Permite resolver
-    // `this.cfg.server.host` em metodos de instancia.
-    {
-        // Stmts a escanear: ctor body (se houver) + init_stmts (initializers
-        // de propriedade que serao inseridos no ctor sintetizado).
-        let mut stmts_to_scan: Vec<&Statement> = Vec::new();
-        for member in &class.members {
-            if let ClassMember::Constructor(ctor) = member {
-                for s in &ctor.body {
-                    stmts_to_scan.push(s);
-                }
-            }
-        }
-        for s in &init_stmts {
-            stmts_to_scan.push(s);
-        }
-        for s in stmts_to_scan {
-            scan_this_obj_assign(
-                s,
-                &mut field_obj_types,
-                &mut field_nested_obj_types,
-            );
-        }
-    }
-
-    let meta = ClassMeta {
-        name: class.name.clone(),
-        super_class: class.super_class.clone(),
-        methods,
-        field_types,
-        field_class_names,
-        field_obj_types,
-        field_nested_obj_types,
-        static_methods,
-        static_fields,
-        getters,
-        setters,
-        has_constructor,
-        readonly_fields,
-        is_abstract: class.is_abstract,
-        abstract_methods,
-        member_visibility,
-        layout: None,
-    };
-    (meta, fns)
-}
-
-/// `this.<name> = <init>;` como Statement RTS.
-fn make_field_init_stmt(
-    name: &str,
-    init: Box<swc_ecma_ast::Expr>,
-    span: crate::parser::span::Span,
-) -> Statement {
-    let lhs = Expr::Member(swc_ecma_ast::MemberExpr {
-        span: Default::default(),
-        obj: Box::new(Expr::This(swc_ecma_ast::ThisExpr {
-            span: Default::default(),
-        })),
-        prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-            span: Default::default(),
-            sym: name.into(),
-        }),
-    });
-    let assign = Expr::Assign(swc_ecma_ast::AssignExpr {
-        span: Default::default(),
-        op: swc_ecma_ast::AssignOp::Assign,
-        left: swc_ecma_ast::AssignTarget::Simple(swc_ecma_ast::SimpleAssignTarget::Member(
-            swc_ecma_ast::MemberExpr {
-                span: Default::default(),
-                obj: Box::new(Expr::This(swc_ecma_ast::ThisExpr {
-                    span: Default::default(),
-                })),
-                prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-                    span: Default::default(),
-                    sym: name.into(),
-                }),
-            },
-        )),
-        right: init,
-    });
-    let _ = lhs; // não usamos; AssignTarget já carrega o lado esquerdo.
-    let stmt = Stmt::Expr(swc_ecma_ast::ExprStmt {
-        span: Default::default(),
-        expr: Box::new(assign),
-    });
-    Statement::Raw(RawStmt::new("<field-init>".to_string(), span).with_stmt(stmt))
-}
-
-/// Costura initializers no body do constructor, respeitando `super()`.
-/// - Se `has_super` e o primeiro statement do user é `super(...)`,
-///   coloca os initializers logo depois.
-/// - Caso contrário, prepende.
-/// (#303 parte 1) Conta \`super(...)\` no nivel top-level do body de um
-/// constructor, sem descer em if/else/loops/blocks. Detect de duplicacao
-/// linear evita o caso degenerate \`super(); super();\`.
-fn count_super_calls_in_top_level(body: &[Statement]) -> usize {
-    use swc_ecma_ast::{Callee, Expr, Stmt};
-    let mut count = 0usize;
-    for stmt in body {
-        let Statement::Raw(raw) = stmt;
-        let Some(s) = raw.stmt.as_ref() else { continue };
-        if let Stmt::Expr(e) = s {
-            if let Expr::Call(c) = e.expr.as_ref() {
-                if matches!(c.callee, Callee::Super(_)) {
-                    count += 1;
-                }
-            }
-        }
-    }
-    count
-}
-
-fn weave_initializers(
-    user_body: &[Statement],
-    init_stmts: &[Statement],
-    has_super: bool,
-) -> Vec<Statement> {
-    if init_stmts.is_empty() {
-        return user_body.to_vec();
-    }
-
-    let mut out: Vec<Statement> = Vec::with_capacity(user_body.len() + init_stmts.len());
-
-    let super_at_start = has_super
-        && user_body
-            .first()
-            .map(|s| is_super_call_stmt(s))
-            .unwrap_or(false);
-
-    if super_at_start {
-        out.push(user_body[0].clone());
-        out.extend(init_stmts.iter().cloned());
-        out.extend(user_body.iter().skip(1).cloned());
-    } else {
-        out.extend(init_stmts.iter().cloned());
-        out.extend(user_body.iter().cloned());
-    }
-
-    out
-}
-
-fn is_super_call_stmt(s: &Statement) -> bool {
-    let Statement::Raw(raw) = s;
-    let Some(Stmt::Expr(expr_stmt)) = raw.stmt.as_ref() else {
-        return false;
-    };
-    let Expr::Call(call) = expr_stmt.expr.as_ref() else {
-        return false;
-    };
-    matches!(call.callee, Callee::Super(_))
-}
-
-/// Body padrão para stub de método abstract: `return 0;` (ou nada se void).
-fn synth_abstract_stub_body(return_type: Option<&str>) -> Vec<Statement> {
-    let ret_type = return_type.map(|s| s.trim()).unwrap_or("void");
-    if ret_type == "void" || ret_type.is_empty() {
-        return Vec::new();
-    }
-    let zero_expr = if ret_type == "f64" || ret_type == "F64" {
-        // f64 → 0.0
-        Expr::Lit(Lit::Num(swc_ecma_ast::Number {
-            span: Default::default(),
-            value: 0.0,
-            raw: None,
-        }))
-    } else {
-        // i32/i64/handle/bool: literal 0
-        Expr::Lit(Lit::Num(swc_ecma_ast::Number {
-            span: Default::default(),
-            value: 0.0,
-            raw: Some("0".into()),
-        }))
-    };
-    let stmt = Stmt::Return(swc_ecma_ast::ReturnStmt {
-        span: Default::default(),
-        arg: Some(Box::new(zero_expr)),
-    });
-    vec![Statement::Raw(
-        RawStmt::new("<abstract-stub>".to_string(), Span::default()).with_stmt(stmt),
-    )]
-}
-
-fn this_param(span: crate::parser::span::Span) -> Parameter {
-    Parameter {
-        name: "this".to_string(),
-        type_annotation: None,
-        modifiers: MemberModifiers::default(),
-        variadic: false,
-        default: None,
-        span,
-    }
-}
-
-pub(super) fn class_init_name(class: &str) -> String {
-    format!("__class_{class}__init")
-}
-
-pub(super) fn class_method_name(class: &str, method: &str) -> String {
-    format!("__class_{class}_{method}")
-}
-
-pub(super) fn class_static_method_name(class: &str, method: &str) -> String {
-    format!("__class_{class}_static_{method}")
-}
-
-pub(super) fn class_getter_name(class: &str, prop: &str) -> String {
-    format!("__class_{class}_get_{prop}")
-}
-
-pub(super) fn class_setter_name(class: &str, prop: &str) -> String {
-    format!("__class_{class}_set_{prop}")
-}
-
